[TNL-6028][TNL-6020] Implement back to beginning button.
The 'Go to Beginning' button is only exposed to keyboard and
screenreader users. It automatically receives focus on next TAB press
after dropping an item, except when a feedback popup is shown after
dropping the item, in which case the focus moves to the 'close' button
of the popup.

Clicking any of these two buttons moves focus back to the first item in
the item bank.