Use sanitized ID for zone descriptions.
In the past, we were using zone title as the zone ID. This is no longer
the case as we programatically generate zone IDs, so this fix is only
relevant to old problems (such as some problems from the DnDv2 demo
course).

We have already been sanitizing zone.uid to be used for zone id
attributes, but we missed zone description id attribute.