Omit info about attempts used when attempts unlimited.
When the number of attempts is unlimited, we were hiding the "You have
used X out of Y attempts" text, which becomes "You have used X out of
null attempts" when attempts are unlimited.

The text was hidden, but it was still read out by the screen reader.
Do not include the text in the DOM when attempts are unlimited at all.